Feature/mainmenu (#10)
* Create the translation menu

It was added the text file translation option in the menu.  The first
option is to translate from english to portuguese and the second option
is to translate from portuguese to english.

* Refacto the operation within de TranslatorManager class to the view
layer

Only the translation part remained inside the TranslatorManager class, everything else went to the view layer, such as choosing the file, saving the translated file and prints.

* Refactor MainMenu to eliminate duplicate print codes

Creating the print function and all methods that need to print calls the print function

* Create a Tkinter window for program output

A Frame was created in Tkinter and within it two widget text boxes were placed, side by side, one for the original text and the other for the translated text. The name of the folder where the files were saved and the processing time are also shown, for those options where this information is required.

* Refactor the code to remove unnecessary imports.

Refactor to remove the imports: from tkinter import filedialog and import os.

* Remove the timeout parameter and include the phrase_time_limit parameter

Remove the timeout parameter and include the phrase_time_limit=10 parameter. The phrase_time_limit parameter is the maximum number of seconds that this will allow a phrase to continue. Whereas, The timeout parameter is the maximum number of seconds that this will wait for a phrase to start. So, as we want to limit the time for the microphone to stop listening, it was necessary to add phrase_time_limit.

---------

Co-authored-by: Yuri Cunha <82489048+Yuricunnha@users.noreply.github.com>
Co-authored-by: João Vitor Silva <vitorpeu@hotmail.com>